Launchpad Travel
This is an update off of Jon002's source, that makes it so you can
travel via rocket to Nimbus Station!

What I did is in WorldLoop.cpp, I mapped the server to to open
"rockettest.bin", and I fixed a glitch in the code that stops the
dreaded 55%/35% glitch!